+++ src/dev-server/notification-server.ts
@@ -1,6 +1,6 @@
 // Ionic Dev Server: Server Side Logger
-import { Logger } from '../util/logger';
-import { hasDiagnostics, getDiagnosticsHtmlContent } from '../util/logger-diagnostics';
+import { Logger } from '../logger/logger';
+import { hasDiagnostics, getDiagnosticsHtmlContent, generateRuntimeDiagnosticContent } from '../logger/logger-diagnostics';
 import { on, EventType } from '../util/events';
 import { Server as WebSocketServer } from 'ws';
 import { ServeConfig } from './serve-config';
@@ -8,6 +8,7 @@ import { ServeConfig } from './serve-config';
 
 export function createNotificationServer(config: ServeConfig) {
   let wsServer: any;
+  const msgToClient: WsMessage[] = [];
 
   // queue up all messages to the client
   function queueMessageSend(msg: WsMessage) {
@@ -75,51 +76,61 @@ export function createNotificationServer(config: ServeConfig) {
     drainMessageQueue();
   });
 
-}
 
+  function printMessageFromClient(msg: WsMessage) {
+    if (msg && msg.data) {
+      switch (msg.category) {
+        case 'console':
+          printConsole(msg);
+          break;
 
-function printMessageFromClient(msg: WsMessage) {
-  if (msg.data) {
-    switch (msg.category) {
-      case 'console':
-        printConsole(msg);
-        break;
-
-      case 'exception':
-        printException(msg);
-        break;
+        case 'runtimeError':
+          handleRuntimeError(msg);
+          break;
+      }
     }
   }
+
+
+  function printConsole(msg: WsMessage) {
+    const args = msg.data;
+    args[0] = `console.${msg.type}: ${args[0]}`;
+
+    switch (msg.type) {
+      case 'error':
+        Logger.error.apply(this, args);
+        break;
+
+      case 'warn':
+        Logger.warn.apply(this, args);
+        break;
+
+      case 'debug':
+        Logger.debug.apply(this, args);
+        break;
+
+      default:
+        Logger.info.apply(this, args);
+        break;
+      }
+  }
+
+
+  function handleRuntimeError(clientMsg: WsMessage) {
+    const msg: WsMessage = {
+      category: 'buildUpdate',
+      type: 'completed',
+      data: {
+        diagnosticsHtml: generateRuntimeDiagnosticContent(config.rootDir,
+                                                          config.buildDir,
+                                                          clientMsg.data.message,
+                                                          clientMsg.data.stack)
+      }
+    };
+    queueMessageSend(msg);
+  }
+
 }
-
-function printConsole(msg: WsMessage) {
-  const args = msg.data;
-  args[0] = `console.${msg.type}: ${args[0]}`;
-
-  switch (msg.type) {
-    case 'error':
-      Logger.error.apply(this, args);
-      break;
-
-    case 'warn':
-      Logger.warn.apply(this, args);
-      break;
-
-    case 'debug':
-      Logger.debug.apply(this, args);
-      break;
-
-    default:
-      Logger.info.apply(this, args);
-      break;
-    }
-}
-
-function printException(msg: WsMessage) {
-
-}
-
-const msgToClient: WsMessage[] = [];
 
 export interface WsMessage {
   category: string;